Keep track of offsets within substreams where possible
Fixes https://github.com/construct/construct/issues/1014

This patch wraps the BytesIO substreams made by the following
subconstructs during parsing:

* OffsettedEnd
* Prefixed
* FixedSize
* NullTerminated
* NullStripped
* ProcessXor

This allows the Tell construct to return offsets that are relative to
the beginning of the stream, rather than relative to the last occurence
of these subconstructs.

This has not been implemented for the following subconstructs:

* Transformed: there is no guarantee that there won't be any byte
  swapping in the transformation, therefore we can't map offsets without
  knowledge of the specific transformation.

* ProcessRotateLeft: tell() offsets would be decreasing, which could
  cause all sorts of problems.

In both those cases, it makes more sense for the user to use a Tell
immediately before the Transformed or ProcessRotateLeft and do their own
mapping.