Add emitfulltype to Optional
I'm in the process of exporting Structs to Kaitai schemas, and the `Optional` field is something I use very often. 

I figure that since `Optional` is typically parsed only when data is available, then using `not _.io.eof` could be used to achieve similar functionality. Or is there something else I'm missing here?